restructure ios
