+++ src/cli/env.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-use crate::io::EnvIO;
+use crate::EnvIO;
 
 #[derive(Clone)]
 pub struct EnvNative {
