+++ src/cli/file.rs
@@ -1,7 +1,7 @@
 use anyhow::Result;
 use tokio::io::{AsyncReadExt, AsyncWriteExt};
 
-use crate::io::FileIO;
+use crate::FileIO;
 
 pub struct NativeFileIO {}
