+++ src/lib.rs
@@ -16,7 +16,6 @@ pub mod grpc;
 pub mod has_headers;
 pub mod helpers;
 pub mod http;
-pub mod io;
 #[cfg(feature = "unsafe-js")]
 pub mod javascript;
 pub mod json;
@@ -26,3 +25,21 @@ pub mod path;
 pub mod print_schema;
 pub mod try_fold;
 pub mod valid;
+
+use std::future::Future;
+
+use http::Response;
+
+pub trait EnvIO: Send + Sync + 'static {
+  fn get(&self, key: &str) -> Option<String>;
+}
+
+#[async_trait::async_trait]
+pub trait HttpIO: Sync + Send + 'static {
+  async fn execute(&self, request: reqwest::Request) -> anyhow::Result<Response<Vec<u8>>>;
+}
+
+pub trait FileIO {
+  fn write<'a>(&'a self, file: &'a str, content: &'a [u8]) -> impl Future<Output = anyhow::Result<()>>;
+  fn read<'a>(&'a self, file_path: &'a str) -> impl Future<Output = anyhow::Result<String>>;
+}
